move editor components to edit domain directory
